first pass at persistent storage within ganache
